Refactor Complete on Test Files
