+++ src/deck.js
@@ -1,0 +1,10 @@
+function createDeck(cards) {
+    let deck = cards
+    return deck
+}
+
+function countCards(deck) {
+    return deck.length
+}
+
+module.exports = {createDeck, countCards}
